+++ rquaidsce/tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(rquaidsce)
+
+test_check("rquaidsce")